add lemon squeezy subscription

+++ app/api/board/route.js
@@ -24,6 +24,13 @@ export async function POST(req) {
 
     const user = await User.findById(session.user.id);
 
+    if (!user.hasAccess) {
+      return NextResponse.json(
+        { error: "You do not have an active subscription. Please subscribe" },
+        { status: 403 }
+      );
+    }
+
     const board = await Board.create({
       userId: user._id,
       name: body.name,
@@ -55,11 +62,18 @@ export async function DELETE(req) {
       return NextResponse.json({ error: "Not authorized" }, { status: 401 });
     }
 
+    const user = await User.findById(session.user.id);
+
+    if (!user.hasAccess) {
+      return NextResponse.json(
+        { error: "You do not have an active subscription. Please subscribe" },
+        { status: 403 }
+      );
+    }
+
     await Board.deleteOne({ _id: boardId }, { userId: session.user.id });
 
-    const user = await User.findById(session.user.id);
     user.boards = user.boards.filter((id) => id.toString() !== boardId);
-
     await user.save();
 
     return NextResponse.json({});
